move decorator to function
