+++ mona_sdk/authentication.py
@@ -29,6 +29,8 @@ from requests.models import Response
 from .logger import get_logger
 from .client_exceptions import MonaAuthenticationException
 
+from functools import wraps
+
 # A new token expires after 22 hours, REFRESH_TOKEN_SAFETY_MARGIN is the safety gap of
 # time to refresh the token before it expires (i.e. - in case
 # REFRESH_TOKEN_SAFETY_MARGIN = 2, and the token is about to expire in 2 hours or less,
@@ -308,3 +310,51 @@ def get_basic_auth_header(api_key, with_auth):
     )
 
 
+class Decorators(object):
+    @classmethod
+    def refresh_token_if_needed(cls, decorated):
+        """
+        This decorator checks if the current client's access token is about to
+        be expired/already expired, and if so, updates to a new one.
+        """
+
+        @wraps(decorated)
+        def inner(*args, **kwargs):
+            # args[0] is the current mona_client instance.
+            mona_client = args[0]
+
+            if not mona_client.should_use_authentication:
+                return decorated(*args, **kwargs)
+
+            # If len(args) < 1, the wrapped function does not have args to log (neither
+            # messages nor config)
+            should_log_args = len(args) > 1 and mona_client.should_log_failed_messages
+
+            # message_to_log is the messages/config that should be logged in case of
+            # an authentication failure.
+            message_to_log = args[1] if should_log_args else None
+
+            if not is_authenticated(mona_client.api_key):
+                return handle_authentications_error(
+                    "Mona's client is not authenticated",
+                    mona_client.raise_authentication_exceptions,
+                    message_to_log,
+                )
+
+            if should_refresh_token(mona_client.api_key):
+                with authentication_lock:
+                    # The inner check is needed to avoid double token refresh.
+                    if should_refresh_token(mona_client.api_key):
+                        refresh_token_response = refresh_token(mona_client)
+                        if not refresh_token_response.ok:
+                            # TODO(anat): Check if the current token is still valid to
+                            #   call the function anyway.
+                            return handle_authentications_error(
+                                f"Could not refresh token: "
+                                f"{refresh_token_response.text}",
+                                mona_client.raise_authentication_exceptions,
+                                message_to_log,
+                            )
+            return decorated(*args, **kwargs)
+
+        return inner
